+++ packages/bridge/src/runtime/vite/client.manifest.mjs
