+++ packages/bridge/src/runtime/vite/vite-node-shared.mjs
