+++ packages/bridge/src/vite/client.ts
@@ -7,13 +7,13 @@ import { getPort } from 'get-port-please'
 import type { ServerOptions, InlineConfig } from 'vite'
 import { defineEventHandler } from 'h3'
 import defu from 'defu'
+import { viteNodePlugin } from '../vite-node'
 import PluginLegacy from './stub-legacy.cjs'
 import { mergeConfig, createServer, build } from './stub-vite.cjs'
 import { devStyleSSRPlugin } from './plugins/dev-ssr-css'
 import { jsxPlugin } from './plugins/jsx'
 import { ViteBuildContext, ViteOptions } from './types'
 import { prepareManifests } from './manifest'
-import { viteNodePlugin } from './vite-node'
 
 export async function buildClient (ctx: ViteBuildContext) {
   const alias = {
